simplify Java union code This example didn't work for Java union foo {   1: BarStruct value } because of naming clash: a general "getValue" method was generated and a specific one for the field "value".
I think a lot of the code in union.java is from Apache Thrift's TUnion APIs. Since we don't implement TUnion, I think we can get rid of them.

In the future though, we'll probably make Scrooge types extend TBase. In that case unions do need to implement TUnion. But this decision has not been made, and if it is, the changes would be across the board.

RB_ID=103926