Include (and optionally compile) external thrift files.
JAR files containing thrift files to include can be added to the list of
library dependencies in the "thrift" configuration. The plugin will unpack the
thrift files from the JARs to an external source directory which is added to
the list of include directories. The external sources can be compiled if the
`scroogeCompileExternalSources` setting is set to `true`.