[split] scrooge-generator: produce slimmer code, remove _passthroughFields from object apply method Attempting to bump the entire birdcage to the current version of scrooge-generator, I ran into a problem where several projects would cause scalac to hit a GC limit.  I've made some tweaks to generate a combination of less code (by removing duplication with read/write methods), code that produces less bytecode (by reducing calls to `clone` which explodes into a lot code with the all the default field values), and code that produces fewer anonymous classes (by removing calls to TFieldBlob.capture).
I have also removed the `_passthroughFields` parameter from the main struct object apply method, because it broke several projects that do reflection or other crazy stuff on that method.  _passthroughFields is present on the Immutable inner-class, which is used internally, so there is no backpedaling on actual functionality.

RB_ID=217990