[split] scrooge: secondary struct class constructors without _passthroughFields for backwards compatibility It was reported to me that enabling _passthroughFields generation by default broke some code in science that uses jackson to deserialize json to thrift.  jackson uses reflection, and passes in null for the _passthroughFields argument, since it doens't know about scala default values.  Consequently, the resulting struct generates NPEs.
To fix this, and any other potential reflection based issues, scrooge now generates two constructors, one with and one without the _passthroughFields parameter.  Scala only allows a single constructor to have default arguments, so I've made it so that the constructor that contains the _passthroughFields parameter has no defaults, since this constructor is primarily intended to be called by the scrooge generated decode method, which passes all constructor args and doesn't depend on defaults.

RB_ID=228563