[split] scrooge: capture unknown union values as its own value (THRIFT-99)
Motivation:
To allow for safely adding new entries in unions. Currently, clients are not forced to deal with unknown types which makes adding and returning new types impossible safely.

Modifications:
Instead of throwing an exception at deserialize time, set the result to an Unknown value. Generated code with pass-through properly maintains pass-through behavior for the unknown union value.

Result:
Clients will now be able to deserialize unions with unknown values, and clients will be able to deal with the unknown value in appropriate way for their use case.

RB_ID=289649