[split] Update scrooge to remove date from Generated annotation so generated code is reproducible.
Currently Scrooge sets the "date" field of the Generated annotation for code
it generates, which leads to unique generated code. This change stops
setting the generated date as its not needed in practice and complicates
pants caching because generated code is unique.

This annotation was originally added in
https://github.com/twitter/scrooge/pull/52 with the stated intent of playing
nice with IDEs. I believe this change still allows the generated code to
play well with editors, as well as now playing well with the pants cache.

RB_ID=290143