[split] scrooge: throw an error when reading a field with the wrong type
Motivation:
Currently Scrooge silently fails when a field is specified but of the wrong type.

Modifications:
Modify readField.scala to throw an exception with more information rather than silently ignoring it.

Result:
When encountering a field with a type different than expected, Scrooge throws an exception that contains the field name as well as the expected and actual types.

RB_ID=293787