[split] finagle-thriftmux: add ClientBuilder#stack compatibility and make APIs symmetric
Problem:

There was no API entry point for ThriftMux into ClientBuilder.

Solution:

Implement (Stack.Params => Client[ThriftClientRequest, Array[Byte]]) on
the ThriftMuxClient. Additionally, the API surface area for thriftmux
was becoming unmanageable. This was partly due to the hoops we had to
jump through to make ThriftMuxServer work with ServerBuilder. However,
now that this is simplified, we have an opportunity to make the API
more symmetric and uniform for both ThriftMuxClient/ThriftMuxServer.
Note, instead of using the StackLike variants we have to define
separate ThriftMux{ServerLike/ClientLike} because of the type
conversions from mux to thrift.

RB_ID=359625
TBR=true