Generated Java code should read/write enums correctly
Problem:
When reading/writing enums over the wire, they are defined
as TType.I32, and not TType.ENUM, as one would expect.
This is arguably a problem with the Thrift spec, but we must
adhere to it anyways.

Solution:
The quirk was already fixed in the generated Scala code.
This ports the enum fix from theScala code generator to
the Java code generator