[split] scrooge-serializer: Remove dependency on scrooge-runtime
Problem

It would be nice to publish scrooge-serializer against 2.11.  For
example, algebird depends on it.

Solution

Because scrooge-serializer only depends on features in scrooge-core,
we can publish properly for scala 2.11 today by changing the
dependency to be on scrooge-core instead of scrooge-serializer.

Result

We can now publish scrooge-serializer against 2.11.

RB_ID=461145