[sbt-plugin] Add thrift files to published artifact
Problem

If someone has a project that compiles thrift but that also depends on a library that defines it's own thrift as well, this project needs to receive the thrift files of the dependency somehow.

Solution

The sbt plugin now adds the thrift source files into the published jar file where both the maven plugin and the sbt plugin are already programmed to look for them.

Result

It is now possible for a project defining it's own thrift to depend on a "thrift" library using the Scrooge sbt plugin to publish it's artifacts.

Fix #151