scrooge: support for ignoring unknown enum ids when reading
Problem:

Adding a value to a thrift enum breaks downstream consumers of the data.
Leading to the ugly "reserved value" pattern.
reading an unknown enum id throws an exception which can not be ignored.
For example we may be filtering on a given known value and ignore everything else.

Solution:

In the case of an unknown enum value provide an instance holding that value,
that will be ignored if we are pattern matching or filtering against some specific values.

Result:

deserializing enum values that are unknown to the consumer (because they have not
deployed with the updated IDL yet) will no longer throw exceptions. Instead they can be ignored.

example:
enum ExternalService {
  FACEBOOK = 1,
  VINE = 2,
  ...
}
say I care only about FACEBOOK and want to ignore everything else
externalService match {
  case FACEBOOK => // do something
  case _ => // don't care, ignore, None
}
the same thing works with filter(_.service == FACEBOOK) or anything that can ignore the unknown ids.
Now if I want to fully understand all the IDs and fail if there is an unknown enum I can do it to:
externalService match {
  case FACEBOOK =>
  ...
  case s: ExternalService => // fail with a message mentioning s.value (matches any non-specified enum, known in the IDL or not)
}
or
externalService match {
  case FACEBOOK =>
  ...
  case EnumUnknownExternalService(id) => // specifically matches only the unknown new ids
}

RB_ID=546478