https://github.com/twitter/scrooge/pull/169
Problem

The gen-test-thrift scripts (in scrooge-generator
and scrooge-benchmark) are OS-specific, so it's
not possible to build Scrooge on Windows. There is
also a lot of code duplication in the scripts.

Solution

Rewrite the scripts into Scala and import them as
sbt tasks. This also gives us the power of Scala.

Signed-off-by: Nik Shkrob <nshkrob@twitter.com>

RB_ID=577046