util/finagle: Small changes to avoid allocations
Problem

In various places we use methods that rely on Scala's implicits to
enrich them. This typically at the cost of an unnecessary object
allocation. There are some other simple changes to be made that
avoid allocations as well.

Solution

Use methods that do not require allocations. Some examples:

* String.nonEmpty => !String.isEmpty
* String.size => String.length
* StringBuilder.size => StringBuilder.length
* Array.size => Array.length
* Seq.empty => Nil

RB_ID=630964