scrooge: make `Immutable` codecs return their "base" codec's metaData
Problem:

For a thrift struct, scrooge will generate an `Immutable` implementation and corresponding companion object/ThriftStructCodec, e.g.:

    object Immutable extends ThriftStructCodec3[UnitTestStruct] {
      override def encode(_item: UnitTestStruct, _oproto: TProtocol) { _item.write(_oproto) }
      override def decode(_iprot: TProtocol): UnitTestStruct = UnitTestStruct.decode(_iprot)
    }

`ThriftStructCodec3`'s interface means this exposes `Immutable.metaData.fields`, which should contain `ThriftStructFields` for all fields of `Immutable`.

The implementation of `fields` uses reflection in a way that returns an empty list for `Immutable`.

Solution:

Override `Immutable`'s `metaData` to return its "base" codec's `metaData` (in the example above, `UnitTestStruct.metaData`.
This is not the minimal change.  The minimal change would be to simply fix the `fields` implementation so that it can find the fields when called on an `Immutable` codec.
The change in this commit will have other effects, since `ThriftStructMetaData` also has vals like `codecClass` and `structName`, which will now return the base class's values instead of `Immutable`'s.  I'm not thoroughly familiar with the ways in which `ThriftStructCodec` is used, but these other effects feel like good things to me.  Then again, it's not clear to me why `Immutable` has a separate `ThriftStructCodec` in the first place, so it's possible I've missed something.

Result:

MyTestStruct.Immutable.metaData.fields returns `ThriftStructFields` for all fields of `MyTestStruct` instead of an empty list.

RB_ID=633376