Update docs with new versions and repository for SBT plugin
Problem

Because we've reverted libthrift to 0.5.0-1, it's now necessary to add the
Twitter Maven repository to the SBT meta-meta-build for projects that use the
Scrooge SBT plugin. The docs don't make this clear.

Solution

Add a line to the example in the docs. I've also updated some other versions
while I'm here.