Provides struct metadata for exception-based fields
Problem:
It is possible to use a thrift exception as field in a thrift struct/union or with a container. The metadata associated with the field (or the container type) says that the field type is a struct, but it is not possible to cast the metadata value (of type FieldValueMetaData) to a struct value metadata (StructMetaData), and to get the class of the type.

Solution:
Uses StructMetaData instead of FieldValueMetaData in the generated code.

Result:
It is now possible to fully introspect a thrift type.

RB_ID=702938