ThriftLinter: Enable TransitivePersistence and DocumentPersisted Rules
Problem

TransitivePersistence and DocumentPersisted rules are currently disabled.

Solution

Enable these rules by default.

Result

Any structs that are annotated with a persisted thrift annotation (persisted =
true) would now be validated. A struct will be invalid if it is annotated
persisted and refers to another struct that is not itself persisted. Currently,
no structs are marked as persisted.

RB_ID=738776