Disable TransitivePersistence and DocumentedPersisted rules
Problem

TransitivePersistence rule needs include-paths to be supplied. That change is in
pants, which is currently pending release. So enabling this causes build to fail
for files the include thrift IDLs from non-local directory. DocumentedPersisted
rule relies on TransitivePersistence rule to work. So that would not work as
expected as well.

Solution

Disabling both rules until the change is pushed in pants.

Result

Build would no longer fail on this with scrooge-gen upgrade.

RB_ID=740624