ThriftLinter: Enable transitive persistence rule
Problem

TransitivePersistence rule is currently disabled.

Solution

Enable the rule by default.

Result

Any structs that are annotated with a persisted thrift annotation (persisted =
true) would now be validated. A strucut will be invalid if it is annotated
persisted and refers to another struct that is not itself persisted. Currently,
no structs are marked as persisted.

RB_ID=748405