scrooge: Fix name conflict in ScroogeSBT*
Problem

SBT plugins are imported into `build.sbt` files with a wildcard import. As such,
SBT plugins should not define public members that collide with members in `sbt._`
and `sbt.Keys._`. `ScroogeSBT` defines a public method `compile` that collides
with `sbt.Keys.compile` and prevents standard usage of `compile` targets in
`build.sbt` files

Solution

Change the scope of the `compile` function to be `private[this]`

Result

Users of the scrooge SBT plugin will be able to properly target the `compile` build
phase to supply specific settings in `build.sbt`.

Signed-off-by: Vladimir Kostyukov <vkostyukov@twitter.com>

RB_ID=760091