[ThriftLinter] Check persisted annotation presence on enum types
Problem

Annotations on enum types were not supported in thrift 0.5 library. Recently, it
was upgraded to 0.9 which supports it. To have complete comverage on presence of
persisted annotation in thrift files, we need to validate it on enum types as
well.

Solution

Add validation check on presence of persisted annotation on enum types.

Result

Moving forward, any enum types being scribed would need to have persisted
annotation.

RB_ID=764491