finagle-thrift,thriftmux: Initial steps towards Thrift ResponseClassifiers
A followup to RB_ID 772906 this is the the beginning of the necessary
wiring for `ResponseClassifiers` for ThriftMux (and later Thrift as
well). This includes the Scrooge changes but it awaits a Scrooge
version bump before it can be fully enabled. This will enable
developers to classify Thrift exceptions or status codes in the
response as failures.

See `c.t.f.thriftmux.service.ThriftMuxResponseClassifier` and
`c.t.f.ThriftMux.withResponseClassifier` to simplify writing and
wiring up classifiers for ThriftMux.

RB_ID=772931