scrooge: Gold file tests for generators
Motivation

It is not always clear to the developer or the reviewer when changes
are made to templates or the generator code what the actual changes
will be to the generated code.

Solution

Introduce golden file tests which capture the current state of the
generated code. Any change to the templates or generators should
manifest as changes in the golden files. These changes can then be
reviewed as part of the change.

RB_ID=781079