scrooge: Use ResponseClassifier for per-method metrics
Motivation

Previous changes wired up user's a custom `ResponseClassifier` to
their Finagle `StatsFilter`. However these classification rules were
not used for Scrooge's per-method metrics.

Solution

Provide the means for the `ResponseClassifier` to be wired through the
`FinagledClient` so that it is available to be used.

Result

It becomes possible for Finagle code to wire this up and that will be
done in a follow up commit (RB 791470). In the meantime, this remains
backwards compatible with prior code as it uses
`ResponseClassifier.Default`.

RB_ID=791465