scrooge: Add originalName to ThriftEnum
Problem

For an enum, generated case objects have an abstract name and value
field, but not originalName.

Solution

Add originalName to the ThriftEnum trait and generated EnumUnknown
case classes. Add default implementation of originalName to return
name to avoid breaking the API.

This is commit 1 of 2. Part 2 will add the changes to ThriftEnum in
scrooge-core.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

RB_ID=814147