scrooge - Fix sbt build for generator used by sbt-plugin in 2.10
Problem

org.scala-lang.modules#scala-parser-combinators should only be
added as a dependency to the scrooge-generator for scala
versions >= 2.11. The generator is used to build the sbt
plugin which is still only scala 2.10 and thus this
dependency is not resovled as there is no 2.10 version.

Solution

Add logic to only add the dependency when the scala version is
>= 2.11.

RB_ID=829998