Allow generic access to ThriftStructCodec3 from instances of ThriftStruct
Please provide a meaningful and concise commit message for
your change. The automatically generated one below is taken from
the ReviewBoard summary and description:

Allow generic access to ThriftStructCodec3 from instances of ThriftStruct

Given an instance of ThriftStruct in a generic context, for example

    def helper[T <: ThriftStruct, U](t): U

There is no good way to get the ThriftStructCodec3 (serializer / deserializer) for "t".
Some codebases resort to reflection here, eg:

    def helper[T <: ThriftStruct, U](t) = {
      val codec = Class.forName(t.getClass.getFullName + "$").newInstance()......
    }

Add a trait that all ThriftStruct's extend from with a method for getting the codec
Now users can:

    def helper[T <: ThriftStruct with HasThriftStructCodec3[T], U](t) = {
      val codec = t._codec
    }

or

    def helper[T <: ThriftStruct, U](t) = {
      val codec = t.asInstanceOf[HasThriftStructCodec3[T]]._codec
    }

This change was initially introduced in RB 830944 and reverted in RB 836953.

RB_ID=868000