scrooge-generator-tests: Fix issues in ServiceGeneratorSpec
Problem/Solution

It looks like we create new listening servers but do not attempt to
close them. Additionally we assert on stats which should be wrapped
in an eventually. These issues are possibly leading to some flakiness
in the ServiceGeneratorSpec. Thus we'll now close the servers and
assert the counters inside a ScalaTest "eventually" block.

RB_ID=873942