Add fieldInfos to ThriftStructMetaData
Please provide a meaningful and concise commit message for
your change. The automatically generated one below is taken from
the ReviewBoard summary and description:

Add fieldInfos to ThriftStructMetaData

Problem
=======

`ThriftStructMetaData` provides detailed information for the fields of union types, but not for non-union types. `unionFields` gives access to the `fieldInfos` companion object method, but `fields` generates a seq of ThriftStructField, which loses some of the information available in the `fieldInfos` method.

Solution
========

Add a `fieldInfos` method to `ThriftStructMetaData` that follows the same pattern as `unionFields`.

RB_ID=874873