scrooge-generator: Fix self referencing type pathological case
Problem

In the case of a defined struct which has one required field
of that struct type, the Java generator effectively creates
two constuctors with the same signatures which does not
compile.

E.g., for IDL defined as such:

struct Hello {
  1: required Hello hello;
}

The Java generator creates a constructor over all *required*
fields and also a "deep copy" constructor over all fields.
In this case, they end up with the same argument list and
thus conflicting and failing to compile.

Solution

When we detect this case we short-circuit to not write out
the "deep copy" constructor and instead users will need to
just call `other.deepCopy` instead.

Result

Compilable generated Java code.

RB_ID=880813