scrooge-generator: Don't allow Structs and Typedefs with the same identifier
Problem

Scrooge currently allows Structs and other types to have the same
identifier. This is a problem when trying to support actual self-
referencing types which necessitates that the identifier is unique
for a given FieldType.

Solution

Revert the work done in RB 880813 in favor of a more correct solution
since it was only a partial fix. We will no longer allow users to
specify a other FieldTypes with the same name as a StructType.

Result

Better support for self-referencing types.

RB_ID=881684