scrooge: Remove check for 22 args when generating scala ServiceIface
Problem

In order to support Scala 2.10 the ServiceTemplate ensured that a case
class was not generated for a ServiceIface when the service defined
more than 22 inherited functions as this was an arity limitation in
Scala 2.10 for case classes.

Solution

Since we no longer support Scala 2.10 we can remove this code
altogether and always generate a case class for the ServiceIface in
Scala.

Result

Less tech debt!

RB_ID=882203