scrooge: Fix TReusableMemoryTransport reallocation on every reset
Problem

TReusableMemoryTransport is used in generated code to reduce byte[]
allocation.

However, due to the way TByteArrayOutputStream is implemented in
libthrift 0.9, buf.reset() will allocate a new underlying byte[] if
current capacity exceeds the initial size (512).

The expected behavior is the underlying byte[] is only reallocated
when current capacity exceeds maxThriftBufferSize, which is what the
above resetBuffer method does.

Solution

Implement a TUnboundedByteArrayOutputStream class that doesn't do any
reallocation and leave that to the resetBuffer method.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

RB_ID=888809