scrooge-generator-tests: Add pants build files
Problem

The vast majority of scrooge-generator-tests do not run in pants. This
has been due to a limitation in pants that required using a published
version of the generator.

Solution

There is now a workaround in place and the code at head can be used as
the generator. Using this, we can run more of the generator unit
tests. I believe with some additional followup work, we can get most
and possibly all of the remaining tests to run.

Result

Most of the Thrift gets generated and compiled in most languages
(Scala, Java, Android).

The test suite runs in about 25 seconds (versus 4 minutes in sbt).
This is 162 tests instead of the full 284 run by sbt, but this is an
improvement over the current 105.

RB_ID=894819