scrooge: Struct unapply returns Option[TupleN]
Problem

Currently, the generated code for thrift structs defines an unapply
method that returns `Option[ProductN]`. This is inconsistent with
Scala's built in behavior when unapply methods are generated for case
classes, which return `Option[TupleN]`. A number of libraries work well
with the latter but not the former (e.g. scala.math.Ordering implicits,
Slick table mappings).

Solution

For consistency with the behavior of Scala case classes, and to improve
library interoperability, this modifies the generated code to return a
tuple rather than product instance. For consistency, the definition of
the ProductN types is also modified to use the `_root_` prefix.

Result

Library interoperability and consistency should be improved; impact on
existing code should be minimal as each TupleN class implements the
corresponding ProductN trait.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

RB_ID=894840