scrooge-generator: Fix default values for collections in scala bindings
Problem:

Generated ThriftStruct classes currently put the wrong value in their
ThriftStructFieldInfo.defaultValue field for fields that are collections.
Currently it puts an empty collection even when there is no configured default
value.

I think this bug comes from the fact that the method that generates
default values is used in 2 different contexts, one for coming
up with the default values for the apply()
method and one for supplying info to the ThriftStructFieldInfo object.

Solution:

Change the generator to put the actual default value, or None if there isn't one
in the ThriftStructFieldInfo.defaultValue field

Result:

The ThriftStructFieldInfo instances in ThriftStructs will have the correct
default value for collection fields (None for fields that have no default, and
Some(default) for fields that do.

RB_ID=908152