scrooge: Fully qualify ProcessFunction in scrooge-generator
Problem

To get everything working on Scala 2.12, we need to fully qualify the ProcessFunction
interface scrooge-generator produces.

Solution

Use "parent_name" as the name of the parent, instead of "name"; "name" is already bound in
the scope of "functions" and thus the parent name is inaccessible without this change. Use
the "parent_name" to qualify the ProcessFunction implementation.

RB_ID=908670