scrooge: Re-enable the ApacheJavaGenerator and CocoaGenerator tests
Problem:

CocoaGenerator tests and one test case in ApacheJavaGenerator are disabled.
ApacheJavaGenerator test fails because the importer does not point to the
correct path, also we need to provide full path of ProcessFunction since we
updated to scala 2.12. CocoaGenerator tests fail because the WordSpec does not
support parameters in test descriptions.

Solution:

Fix and re-enable them.

RB_ID=914031