scrooge: Fix ApacheJavaGeneratorSpec Thrift Import
Problem

When running scrooge-generator-tests from SBT, it is possible to wind up
with incorrect file paths for the Importer if trying to load files from
the file system via relative paths. This happens for one test in
ApacheJavaGeneratorSpec.

Solution

Create a TestThriftResourceImporter which takes a filename and loads the
FileContents via resource loading rather than trying to correctly
determine the full path on the file system.

RB_ID=914903