finagle-mux: Removed the dependency on commons stats
Summary: Problem

finagle-mux depends on commons stats to compute latency quantiles.  It then uses
that to decide how long to give clients in leases.  However, it only uses a
single class from commons stats, and commons stats is a relatively heavy
dependency, so it would be better to avoid it if possible.

Solution

Switch from WindowedApproxHistogram to LatencyHistogram, which lives in
finagle-exp.

Result

finagle-mux no longer depends on commons.

JIRA Issues: CSL-4665

Differential Revision: https://phabricator.twitter.biz/D55598