Adaptive Scrooge - deserialize only used fields of thrift Adaptive Scrooge - deserialize only used fields of thrift
Problem

Many times a service needs to get only a small subset of data from that
provided by the available api of another service or stream (e.g Kafka
stream). The binary serializer parses the bytes to materialize the entire
nested tree of objects. Lazy serializer does better by parsing strings and
optional primitives lazily but it still creates the entire nested tree
of thrift objects. It will be great to be able to avoid deserializing
entire trees.

Solution

Adaptive Scrooge observes which fields of a thrift struct are accessed
and based on that adapts the decoding logic at run time to avoid creating
objects for unused fields. This way we can eliminate object creation
for entire trees.

I've implemented Adaptive Scrooge. I've updated the code generator to
generate relevant code and implemented helper classes. I've provided
tests and benchmarks as well. I'm planning to add more benchmarks.

Implementation notes:

The only way of communicating with the generated code is via the protocol
so the design uses it to customize Adaptive Scrooge decoding. e.g. the
protocol provides methods for building the decoder that generated code
uses. This way settings like number of events to learn from and number of
accesses to consider a field used can be kept outside the generated code
and are thus customizable.  Amount of modifications done to the classes
via ASM is kept to a minimum, code is generated to help ASM. e.g. we
use methods with special names whose sole purpose is to indicate to
ASM the sections it may need to prune. This is done to keep debugging
easier, e.g. stack traces can be meaningfully correlated with real code.
The classes that are modified via ASM are kept outside the companion
object, so that they don't carry any reference to it. Due to this reason
some private methods in the companion object have been made package
private for reuse.  Special methods added to some classes are formatted
to avoid possible conflicts to other names that may naturally be present
in the thrift. e.g. member variables in Adapt classes are prefixed with
m_.

Result

Improvements depend on number of fields accessed. I've provided a
benchmark where a tiny portion of thrift object is accessed and there
Adaptive Scrooge is 10x faster than Lazy Scrooge, but it's not a common
case. I've also tried with some real production events with real access
patterns and there Adaptive Scrooge was 1.5x to 2x faster than Lazy
Scrooge. I also ran a real summingbird job using this optimization and
saw the cpu usage of the job drop by 10%. Benefits for different jobs
will of course vary a lot depending on how much overall time they're
spending on serialization and their access patterns.

As a result of this change more code will be generated for scrooge
 scala so there is some compilation time hit. There are
no additional library dependencies if user doesn't make use of Adaptive
Scrooge. To use Adaptive Scrooge user will need to depend on a new
library I've added which among other things also depends on ASM. Users
will need a special serializer to use Adaptive Scrooge. Other than that
they shouldn't need to change any code, Adaptive Scrooge is a drop in
replacement. There will be some initial performance hit when classes
are modified using ASM and loaded into the jvm. For the cases I've
tried this time is very short, less than couple hundred milliseconds.
Another effect of AdaptiveScrooge is that every part of the thrift object
keeps a reference to the original byte array. So in case a field that
we considered unaccessed is accessed we can fallback to decoding from the
original bytes. So if such an object is stored in a cache the whole byte
blobs will stay with them. Users will have to be careful to call
copy on the objects when they store them in a cache.

RB_ID=908416