scrooge-generator: reduce warnings and support modules for Cocoa
Summary: Problem

The Objective C code generated from the Scrooge Cocoa generator results
in compiler warnings with certain inputs. In addition, the public header
files do not support putting the code in a modular framework.

Solution

1. Loop iterators and parameters to initWithCapacity now explicitly
convert from a signed integer to an unsigned one. This prevents warnings
and integer overflows.

2. The NSDecoder methods now explicitly cast the result to the expected
type. This is largely unnecessary, though not harmful, except in the
case of the Thrift type i16, where an implicit cast from a 32-bit
integer to the 16-bit member var results in a warning. NSDecoder only
provides a decodeInt32ForKey: method, and not a decodeInt16ForKey:.

3. The generated header files that need to reference other generated
header files now use a modular import syntax. That is, instead of:

#import "Filename.h"

it now generates:

#import <Namespace/Filename.h>

4. ApacheThrift is now expected to be a module and uses the "@import
ApacheThrift" syntax.

JIRA Issues: FAB-1590

Differential Revision: https://phabricator.twitter.biz/D74200