[strict-deps] Mark remaining CSL targets as strict_deps=True. Add exports to CSL targets
Summary: Problem

Not all CSL targets use strict_deps. Many targets that depend on CSL targets implicitly rely on those targets dependencies being on the compile classpath.

Solution

This changeset converts the remaining core targets to strict_deps, and adds exports to the targets that expose types from their dependencies. For the most part this was done using a script that analyzed the used types in the compiled class files.

The export generation script generated exports for dependencies that are necessary to compile against the types within a given target. It did that by collecting the dependencies that owned super types or annotations used in each target.

Result

All CSL targets have strict_deps enabled. For CSL targets that have annotations or super types that may be necessary for their dependent targets to compile, they export their dependencies that provide those annotations and types.

Things to note:
* There are some cases where it appears to have removed dependencies. This is because there are places where the script would prune dependencies that were provided by an alias.
* The script doesn't take into account pickled Scala types, so the cases where they were the source of compilation failures were handled manually.
* Sorting. The script sorted the dependency lists and the export lists. In the cases where the lists were already sorted, that wasn't a large change, but some targets dependencies were only partially sorted before.

Differential Revision: https://phabricator.twitter.biz/D80632