Twitter OSS: Consistently skip flaky tests in projects
Summary: Problem

The build.sbt files for util, scrooge, finagle, twitter-server, and
finatra do not consistently set the SKIP_FLAKY system
property by default for tests. Additionally, the travis configuration
for said projects is inconsistent, making it hard to spot potential
issues.

Solution

Try to fork when testing and pass the SKIP_FLAKY paramater
to the forked testing process. Every project already specified
that parallel test running in the forked process should be disabled.
However in the case of Finagle this was not enough, so an
additional restriction was added (like in Finatra) to ensure that
tests are not run in parallel in the forked process since test
state will collide.

Result

Tests marked as flaky will be skipped when building with sbt.

JIRA Issues: CSL-5017

Differential Revision: https://phabricator.twitter.biz/D86877