[scala-strict-deps] Remove strict_deps=True for targets between notificationservice/api/src/main/scala/com/twitter/notificationservice/api/BUILD and src/scala/com/twitter/ads/common/base/BUILD
Summary: This is an automated change. We’re not expecting you to review the individual changes to the BUILD files you own, but we want you to be aware that they are happening.

This patch removes strict_deps=True for a partition of the Scala targets in Source because it is enabled by default globally. The changes were generated by a script locally that is very similar to the CI job, https://ci.twitter.biz/job/enable_strict_deps_scala that initially enabled strict_deps.

The script only removes the explicit setting of `strict_deps=True` in BUILD files, and does not change any other aspect of the build file. For more information about strict_deps, please refer to the [[http://go/strict-deps-scala-rollout | Strict Deps Implementation]] doc.

Strict deps restricts what is on the compile classpath to improve compile times. This means you spend less time waiting for your builds. However, with this change, you need to remember that transitive dependencies don't work so if you need something to compile code, you need to add it directly to your dependencies.

The targets affected are Scala targets with strict_deps enabled and names between
  - notificationservice/api/src/main/scala/com/twitter/notificationservice/api/BUILD
  - src/scala/com/twitter/ads/common/base/BUILD

TBR=true

Differential Revision: https://phabricator.twitter.biz/D96903