scrooge-generator: Add relativePath method to GoldFileTest
Summary: Problem

As part of one of the `GoldFileTest`s, it is necessary to construct a
Seq[(String, Buf)] in order to check that the types output by the
generator are correct and that all external references can be resolved.
In order to do this, we need to relativize the path of each of the
generatedFiles against the root of the output (i.e. the temp directory).

Additionally, two small usability fixes have been included.

The temp directory name generated by scrooge gives no indication that it
was generated by scrooge. Change it so that 'scrooge' appears in the
name for easier identification in $TMPDIR.

`ScroogeInternalException` does not currently take the `cause` param as
part of its constructor, which prevents one from throwing it with an
underlying exception instance.

Solution

In order to provide a relative path, I've added a relativePath method
that takes the `java.io.File` object from the generatedFiles list and
relativizes it against the current tempdir. If the `File` object passed
is not rooted in tempdir, an exception is thrown. Compared with
implementing a getter for the tempdir directly, I thought this was the
more conservative approach.

I also changed the generated tempdir name from `"temp${RANDOM}dir"` to
`"scrooge${RANDOM}temp"` to make it easier to identify in `ls` output.

Change the constructor of `ScroogeInternalException` to allow one to
specify the underlying root cause of the exception.

Result

GoldFileTest now has a relativePath method, creates a tempdir that
has 'scrooge' in the name, and `ScroogeInternalException` allows one
to specify a cause.

Differential Revision: https://phabricator.twitter.biz/D98324