finagle-thrift, thriftmux, scrooge: Move Tracing of RPC method names
Summary: Problem

The tracing of Thrift method names happens in the Finagle layer which
does not have easy access to the method being called. In order to do
tracing at this layer, part of the Thrift message must be decoded,
which is a waste of resources.

Solution

Push the tracing down into the generated code where the method name is
known.

JIRA Issues: csl-5577

Differential Revision: https://phabricator.twitter.biz/D115294