scrooge-generator: Remove ThriftResponse.firstException usage in generated server code
Summary: Problem:

ThriftResponse.firstException is calling `collectFirst`. This is too expensive
because we know only one exception field in the result is set.

Solution:

Changed ThriftExceptionResponse to only take the exception instead of the
result, and renamed subclasses of RichResponse to indicate they are responses,
not necessarily contain results.

JIRA Issues: CSL-5686

Differential Revision: https://phabricator.twitter.biz/D132202