scrooge-generator: Implement asClosable on ReqRep/ServicePerEndpoint
Summary: Problem

We generate an `asClosable` function on the higher-kinded interface
and is thus inherited by the `MethodPerEndpoint` interface. The
base implementation of this method produces a no-op, however the
`MethodPerEndpointImpl` which is generated over a `ServicePerEndpoint`
provides an implementation which closes all the service methods of
the `ServicePerEndpoint`. The same is true for the ReqRep case.

However, there is no way to close a `ServicePerEndpoint` or
`RepRepServicePerEndpoint` client interface directly as the current
functionality is only on the higher-kinded type and it's subclasses.

Solution

Update to add `asClosable` to both the `ServicePerEndpoint` and
`RepRepServicePerEndpoint` interfaces with full implementations in
their respective implementations. The current `MethodPerEndpointImpl`
implementations merely proxy to the `ServicePerEndpoint` version.

JIRA Issues: CSL-5943

Differential Revision: https://phabricator.twitter.biz/D134171