scrooge-generator: Add support for recursive structs
Summary: Problem

Currently, scrooge generator is not able to work with mutually recursive structs
of the form:

```
struct A {
  1: optional B b
}

struct B {
  1: optional A a
}
```

This issue is tracked here: https://github.com/twitter/scrooge/issues/160

Solution

Thrift currently does not support nested structs (and does not plan to do so).
We're taking advantage of this fact: before resolving each field in a thrift
doc, we're adding all the top-level structs to a temporary dictionary. We're
then using that dictionary when resolving a field fails because it's a recursive
definition, from further in the file.

Differential Revision: https://phabricator.twitter.biz/D134470