[Scrooge] add support for construction required annotation
Summary: **Problem**
There are reasons to add optional fields to a thrift definition, but to want the
field to be required for new instance creation. Ideally, we could have the same
compile and run time checks that a field is populated for these fields.

**Solution**
Add a new annotation "constructor_required" to indicate that a field is required
at construction time. This annotation should be used by the compiler to change
all the public constructors for the struct that contain the field. The
constructors should not take an option for these fields.

Also add a run time validateNewInstance check for the presense of these fields.
This is most important for validating thrift objects which have been
deserialized.

There are three public constructors for structs: the apply method and two
Immutable constructors. I modify the apply method on the struct object. I make
the existing Immutable constuctors package private and add two new public ones
that match the API that we want.

The copy method is another public API for that I change. The existing copy API
will change the value if the value is provided. So if None is provided the value
will change to not be present. We don't want this to be possible. If we changed
the API of copy to accept get rid of the Option for the field, there is no way
to keep the value as unset if it was already unset (say through
deserialization). So the existing API does not work. I added a new copy method
with different sematics for the construction required methods. The method allows
us to keep the value as is (by providing the None value, default) or provide a
new value (by providing a Some).

JIRA Issues: AVM-6682

Differential Revision: https://phabricator.twitter.biz/D148841