[Scrooge] Make tests more strict. Document how to generate test data
Summary: **Problem**
Some of the tests in the scrooge code base that compare gold data to generated
data are lax. They ignore comments, whitespace, and even some actual code
differences. This can lead to problems where reviewers cannot tell if changes to
gold data are related to changes or not.

**Solution**
Make the tests much more strict.

When making the tests stricter, I also documented how to generate the gold data
for the various tests. I thought this was necessary, so that contributors don't
need to insert white space manually to make tests succeed with the new
restrictions. I don't give the exact scrooge commands to run, because they are
different for pants and sbt. The arguments should be the same though.

I modified two things about how the apache and java generator tests worked to
make them compatible with the scrooge CLI. I always generate the hashCode and I
set a default package name of thrift. I think that things were setup this way
for easier testing and not to test any feature. The ability to turn off hashCode
generation is not a publicly available feature so I don't think it needs to be
tested. And the default package for the CLI is thrift so I did not think the
tests was intentionally testing it's omission.

I left in the AndroidGeneratorSpec and ApacheJavaGeneratorSpec tests that
generate gold data. If I owned this project, I might take them out. They test
the same thing as the GoldFileTests and generating the gold files for them is
more complicated.

Differential Revision: https://phabricator.twitter.biz/D155366