[Scrooge] add support for construction required annotation in java
Summary: **Problem**
Support for the construction required annotation does not exist in
java.

**Solution**
Add support for the annotation.

The constructor that that accepts only required fields should accept
all fields make as constuction required. Also add a static
validateNewInstance method to generated thrift structures. It returns
a list of strings which describe the ways in which an object does not
validate. The validation issues are:
- Missing a required field
- Missing a construction required field
- Missing a set value in a union struct

I have it return strings rather than typed ValidationIssue like we do
in scala, because I did not want to add a dependency to the objects
that are generated.

I refactored the DeepCopyController so that it can be used for both
deep copy and deep validation.

JIRA Issues: AVM-6882

Differential Revision: https://phabricator.twitter.biz/D151861